add docs for rotate90
